+++ src/index.tsx
@@ -1,4 +1,6 @@
 export { default as State } from './State';
+export { default as Toggle } from './Toggle';
+export { default as Clipboard } from './Clipboard';
 export { default as Effect } from './Effect';
 export { default as Memo } from './Memo';
 export { default as Portal } from './Portal';