+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 import { CProps } from '../../types';
-import { ElementType, Fragment, ReactNode, createElement } from 'react';
+import { Fragment, ReactNode, createElement } from 'react';
 
 export function element({ as, children, ...rest }: CProps & { children: ReactNode }) {
   return createElement(as || Fragment, rest, children);